+++ lib/src/actions.dart
@@ -29,6 +29,8 @@ class CustomSlidableAction extends StatelessWidget {
     this.backgroundColor = _kBackgroundColor,
     this.foregroundColor,
     this.autoClose = _kAutoClose,
+    this.borderRadius = BorderRadius.zero,
+    this.padding,
     required this.onPressed,
     required this.child,
   })  : assert(flex > 0),
@@ -72,6 +74,18 @@ class CustomSlidableAction extends StatelessWidget {
   /// {@endtemplate}
   final SlidableActionCallback? onPressed;
 
+  /// {@template slidable.actions.borderRadius}
+  /// The borderRadius of this action
+  ///
+  /// Defaults to [BorderRadius.zero].
+  /// {@endtemplate}
+  final BorderRadius borderRadius;
+
+  /// {@template slidable.actions.padding}
+  /// The padding of the OutlinedButton
+  /// {@endtemplate}
+  final EdgeInsets? padding;
+
   /// Typically the action's icon or label.
   final Widget child;
 
@@ -89,10 +103,13 @@ class CustomSlidableAction extends StatelessWidget {
         child: OutlinedButton(
           onPressed: () => _handleTap(context),
           style: OutlinedButton.styleFrom(
+            padding: padding,
             backgroundColor: backgroundColor,
             primary: effectiveForegroundColor,
             onSurface: effectiveForegroundColor,
-            shape: const RoundedRectangleBorder(),
+            shape: RoundedRectangleBorder(
+              borderRadius: borderRadius,
+            ),
             side: BorderSide.none,
           ),
           child: child,
@@ -130,6 +147,8 @@ class SlidableAction extends StatelessWidget {
     this.labelPosition = _labelPosition,
     this.spacing = 4,
     this.label,
+    this.borderRadius = BorderRadius.zero,
+    this.padding,
   })  : assert(flex > 0),
         assert(icon != null || label != null),
         super(key: key);
@@ -162,6 +181,12 @@ class SlidableAction extends StatelessWidget {
 
   /// A label to display below the [icon].
   final String? label;
+
+  /// Padding of the OutlinedButton
+  final BorderRadius borderRadius;
+
+  /// Padding of the OutlinedButton
+  final EdgeInsets? padding;
 
   @override
   Widget build(BuildContext context) {
@@ -224,6 +249,8 @@ class SlidableAction extends StatelessWidget {
           );
 
     return CustomSlidableAction(
+      borderRadius: borderRadius,
+      padding: padding,
       onPressed: onPressed,
       autoClose: autoClose,
       backgroundColor: backgroundColor,